Add PATH/LD_LIBRARY_PATH setup to .bashrc as well
The MPI test scripts were using the EFA installer libfabric on the remote node
and the built libfabric on the node where the test is invoked.

It looks like the ldconfig bits are not being loaded for non-interactive
shells. And it's safer to just set both as it varies on distros and the type of
shell invoked.

A follow on patch will pass LD_LIBRARY_PATH for OMPI mpirun.