A script to test the -m option of EFA Installer
EFA Installer now has an -m (--minimal) option to install kenerl driver
and rdma-core only (not installing libfabric and openmpi), this is
designed for users who want only to use Intel MPI, which comes with
its own version of libfabric.

This script test the -m option by installing it then run MPI tests
(ring_c and omb). This script does not run fabtests because the
libfabric comes with Intel MPI does not have header file, and fabtests
cannot be test against it.

Signed-off-by: Wei Zhang <wzam@amazon.com>