Use SSH connection test to check if the instance is ready
We used to check the instance status by calling
"aws ec2 describe-instance-status" in a loop, and set the
loop count to 100. But the check sometime still failed with
"insufficient data".

Use SSH connection test to check if the instance is ready,
instead. It retries with exponential backoff.
The initial backoff is 30s, and doubles for each retry,
until the maximum (16 minutes).

Signed-off-by: Jie Zhang <zhngaj@amazon.com>