Only do huge page number check when the huge page size is not greater than 16 MB
RHEL 8 on ARM has a different huge page configuration than other
OSes and defaulted to 512 MB huge page size. This might
1. massively over-reserve huge pages, thus run the machine effectively OOM
2. not reserve sufficient pages for a process per core worth of huge pages
To deal with this issue, EFA-Config package sets a huge page size threshold as 16 MB (16384 KB).
If the default huge page size is larger than the threshold, the huge page is not used.

Signed-off-by: Shi Jin <sjina@amazon.com>