Reduce the retry number and sleep time for create_instance function.
Currently, create_instance() retry 30 times to create instances, and
sleep 2 minutes between each retry. This could cost totally 2*30=60
minutes to create instances, which is too long. This patch reduces the
number of retry to 10 and the sleep time to 1 minutes. So the total
time cost of instance creation will not exceed 10 minutes.

Signed-off-by: Shi Jin <sjina@amazon.com>