Check if GPUDirect RDMA is supported by reading from sysfs file
Currently, the efa-check.sh check if GPUDirect RDMA is supported
by checking if gdr: Y is present in efa's modinfo. This approach
does not apply to the latest efa kernel driver (1.14.1) which combines
the gdr/non-gdr packaging and has this line present in its modinfo always.
The correct way to check if gdr is enabled is to read the value of sysfs file
"/sys/class/infiniband/<device_name>/gdr", which should be 1 when p2pmem driver
is available and 0 otherwise.

Signed-off-by: Shi Jin <sjina@amazon.com>