Disable running fabtests with tcp provider on ARM instance
EFA kernel module with ARM architecture is not shipped in the installer
for CentOS/RHEL 7. Currently, we are using a1.4xlarge instance and
running fabtests with libfabric tcp provider in centos7-arm and rhel7-arm
(Skip-Kmod test) in EFAInstallerProdCanary and LibfabricMainCanary.

However, we are seeing random failures with tcp provider, which causes
EFAInstallerProdCanary and LibfabricMainCanary jobs unstable. Therefore,
disable running fabtests with tcp provider in centos7-arm and rhel7-arm
skip-kmod test.

Signed-off-by: zhngaj <zhngaj@amazon.com>